+++ src/main/java/umc/plantory/domain/diary/service/DiaryCommandUseCase.java
@@ -1,0 +1,8 @@
+package umc.plantory.domain.diary.service;
+
+import umc.plantory.domain.diary.dto.request.DiaryRequestDTO;
+import umc.plantory.domain.diary.dto.response.DiaryResponseDTO;
+
+public interface DiaryCommandUseCase {
+    DiaryResponseDTO.DiaryInfoDTO saveDiary(DiaryRequestDTO.DiaryUploadDTO request);
+}
